Change CORS, Dto's and type data

+++ src/main/java/com/consorcio/servicios/Security/Dto/RegisterRequestDto.java
@@ -21,6 +21,6 @@ public class RegisterRequestDto {
     @NotNull
     int dni;
     @NotNull
-    int phone;
+    long phone;
 
 }
